requirements.txt
fastapi==0.104.1
openai==1.3.5
uvicorn==0.24.0
python-dotenv==1.0.0
python-telegram-bot==20.6
requests==2.31.0
bcrypt==4.1.2
email-validator==2.1.0